+++ resources/views/components/tabs.blade.php
@@ -7,16 +7,53 @@
     $persistKey = is_string($persist) && $persist !== '' ? $persist : null;
     $tabList = is_array($tabs) ? $tabs : [];
     $defaultId = $tabList[0]['id'] ?? '0';
-    $storageKey = $persistKey !== null ? 'vellum-tabs-'.$persistKey : null;
-    if ($storageKey !== null) {
-        $xData = '{ active: (typeof localStorage !== "undefined" && localStorage.getItem("'.$storageKey.'")) || "'.$defaultId.'", set(id) { this.active = id; localStorage.setItem("'.$storageKey.'", id) } }';
-    } else {
-        $xData = '{ active: "'.$defaultId.'", set(id) { this.active = id } }';
-    }
     $class = $code ? 'vellum-tabs vellum-tabs-code' : 'vellum-tabs';
+    static $groupSeq = 0;
+    $groupId = 'vt'.(++$groupSeq);
 @endphp
 @if ($tabList === [])
 <div class="{{ $class }}" data-vellum-tabs=""@if ($persistKey) data-persist="{{ $persistKey }}"@endif>{!! $slot !!}</div>
 @else
-<div class="{{ $class }}" data-vellum-tabs="" x-data="{{ $xData }}"@if ($persistKey) data-persist="{{ $persistKey }}"@endif><div class="vellum-tabs-list" role="tablist">@foreach ($tabList as $tab)<button type="button" class="vellum-tabs-trigger" role="tab" :aria-selected="active === '{{ $tab['id'] }}'" @@click="set('{{ $tab['id'] }}')" id="vellum-tab-{{ $tab['id'] }}">{{ $tab['label'] }}</button>@endforeach</div>@foreach ($tabList as $index => $tab)<div class="vellum-tabs-panel" role="tabpanel" x-show="active === '{{ $tab['id'] }}'" aria-labelledby="vellum-tab-{{ $tab['id'] }}"@if ($index !== 0 || $storageKey !== null) x-cloak=""@endif>{!! $tab['html'] !!}</div>@endforeach</div>
+<div
+    class="{{ $class }}"
+    data-vellum-tabs
+    x-data="vellumTabs(@js($defaultId), @js($persistKey))"
+    @if ($persistKey) data-persist="{{ $persistKey }}" @endif
+>
+    <div class="vellum-tabs-list" role="tablist" aria-orientation="horizontal" x-on:keydown="onListKeydown($event)">
+        @foreach ($tabList as $tab)
+            @php
+                $tabDomId = $groupId.'-tab-'.$tab['id'];
+                $panelDomId = $groupId.'-panel-'.$tab['id'];
+            @endphp
+            <button
+                type="button"
+                class="vellum-tabs-trigger"
+                role="tab"
+                id="{{ $tabDomId }}"
+                data-value="{{ $tab['id'] }}"
+                aria-controls="{{ $panelDomId }}"
+                :aria-selected="(active === @js($tab['id'])).toString()"
+                :tabindex="active === @js($tab['id']) ? 0 : -1"
+                x-on:click="select(@js($tab['id']))"
+            >{{ $tab['label'] }}</button>
+        @endforeach
+    </div>
+    @foreach ($tabList as $index => $tab)
+        @php
+            $tabDomId = $groupId.'-tab-'.$tab['id'];
+            $panelDomId = $groupId.'-panel-'.$tab['id'];
+        @endphp
+        <div
+            class="vellum-tabs-panel"
+            role="tabpanel"
+            id="{{ $panelDomId }}"
+            data-value="{{ $tab['id'] }}"
+            aria-labelledby="{{ $tabDomId }}"
+            :aria-hidden="(active !== @js($tab['id'])).toString()"
+            x-show="active === @js($tab['id'])"
+            @if ($index !== 0) x-cloak @endif
+        >{!! $tab['html'] !!}</div>
+    @endforeach
+</div>
 @endif